Move QueryTabList to shared dir
